requirements.txt
python-dotenv==1.0.0
psycopg2-binary==2.9.6
pandas==2.0.0
matplotlib==3.7.1
seaborn==0.12.2
tabulate==0.9.0
loguru==0.7.0
requests==2.28.2
notebook==6.5.4
numpy==1.24.3